+++ src/main/java/roomescape/service/ReservationService.java
@@ -1,17 +1,12 @@
 package roomescape.service;
 
 import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.LocalTime;
-import java.time.temporal.ChronoUnit;
 import java.util.Comparator;
 import java.util.List;
-import java.util.Optional;
 import org.springframework.stereotype.Service;
-import roomescape.exception.BadRequestException;
-import roomescape.exception.DuplicatedException;
 import roomescape.exception.NotFoundException;
 import roomescape.model.Reservation;
+import roomescape.model.ReservationSavePolicy;
 import roomescape.model.ReservationTime;
 import roomescape.model.Waiting;
 import roomescape.model.member.LoginMember;
@@ -27,13 +22,16 @@ public class ReservationService {
     private final ReservationRepository reservationRepository;
     private final ReservationTimeRepository reservationTimeRepository;
     private final WaitingRepository waitingRepository;
+    private final ReservationSavePolicy reservationSavePolicy;
 
     public ReservationService(ReservationRepository reservationRepository,
                               ReservationTimeRepository reservationTimeRepository,
-                              WaitingRepository waitingRepository) {
+                              WaitingRepository waitingRepository,
+                              ReservationSavePolicy reservationSavePolicy) {
         this.reservationRepository = reservationRepository;
         this.reservationTimeRepository = reservationTimeRepository;
         this.waitingRepository = waitingRepository;
+        this.reservationSavePolicy = reservationSavePolicy;
     }
 
     public List<Reservation> findAllReservations() {
@@ -41,11 +39,7 @@ public class ReservationService {
     }
 
     public Reservation saveReservation(ReservationDto reservationDto) {
-        ReservationTime time = findReservationTime(reservationDto);
-
-        LocalDate date = reservationDto.getDate();
-        validateIsFuture(date, time.getStartAt());
-        validateDuplication(date, time.getId(), reservationDto.getThemeId());
+        reservationSavePolicy.validate(reservationDto);
 
         Reservation reservation = new Reservation(reservationDto);
         return reservationRepository.save(reservation);
@@ -75,27 +69,6 @@ public class ReservationService {
 
     public List<Reservation> findReservationsByMember(LoginMember member) {
         return reservationRepository.findByMemberId(member.getId());
-    }
-
-    private ReservationTime findReservationTime(ReservationDto reservationDto) {
-        long timeId = reservationDto.getTimeId();
-        Optional<ReservationTime> time = reservationTimeRepository.findById(timeId);
-        return time.orElseThrow(() -> new NotFoundException("[ERROR] 존재하지 않는 데이터입니다."));
-    }
-
-    private void validateIsFuture(LocalDate date, LocalTime time) {
-        LocalDateTime timeToBook = LocalDateTime.of(date, time).truncatedTo(ChronoUnit.SECONDS);
-        LocalDateTime now = LocalDateTime.now().truncatedTo(ChronoUnit.SECONDS);
-        if (timeToBook.isBefore(now)) {
-            throw new BadRequestException("[ERROR] 현재 이전 예약은 할 수 없습니다.");
-        }
-    }
-
-    private void validateDuplication(LocalDate date, long timeId, long themeId) {
-        boolean isExist = reservationRepository.existsByDateAndTimeIdAndThemeId(date, timeId, themeId);
-        if (isExist) {
-            throw new DuplicatedException("[ERROR] 중복되는 예약은 추가할 수 없습니다.");
-        }
     }
 
     private void updateWaitingToReservation(Reservation reservation) {
